backend/requirements.txt
pdfplumber
sentence-transformers
chromadb
faiss-cpu
uvicorn
fastapi
pydantic
numpy